Update .gitignore
R Code for Structural Equation Modeling using Partial Least Squares (PLS-SEM)

This repository contains the RStudio code used for Structural Equation Modeling via Partial Least Squares (PLS-SEM) applied to the collected data of a Theory of planned behaviour and discrete choice experiment survey. The code has been specifically developed for this study and allows for the replication of the results that will be published.